+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/DefaultAudioVideoController.kt
@@ -23,12 +23,17 @@ class DefaultAudioVideoController(
     private val videoClientObserver: VideoClientObserver
 ) : AudioVideoControllerFacade {
     override fun start() {
+        start(AudioVideoConfiguration())
+    }
+
+    override fun start(audioVideoConfiguration: AudioVideoConfiguration) {
         audioClientController.start(
-            configuration.urls.audioFallbackURL,
-            configuration.urls.audioHostURL,
-            configuration.meetingId,
-            configuration.credentials.attendeeId,
-            configuration.credentials.joinToken
+            audioFallbackUrl = configuration.urls.audioFallbackURL,
+            audioHostUrl = configuration.urls.audioHostURL,
+            meetingId = configuration.meetingId,
+            attendeeId = configuration.credentials.attendeeId,
+            joinToken = configuration.credentials.joinToken,
+            audioMode = audioVideoConfiguration.audioMode
         )
         videoClientController.start()
     }