+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/fragment/MeetingFragment.kt
@@ -28,6 +28,7 @@ import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
+import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DiffUtil
@@ -38,6 +39,7 @@ import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributes
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.analytics.toJsonString
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AttendeeInfo
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoConfiguration
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoFacade
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoObserver
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.SignalUpdate
@@ -47,6 +49,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptItem
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptionStatus
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptionStatusType
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.VolumeUpdate
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioMode
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerdetector.ActiveSpeakerObserver
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerpolicy.DefaultActiveSpeakerPolicy
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareObserver
@@ -65,6 +68,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.gl.EglCoreFact
 import com.amazonaws.services.chime.sdk.meetings.device.DeviceChangeObserver
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDevice
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDeviceType
+import com.amazonaws.services.chime.sdk.meetings.internal.AttendeeStatus
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.DefaultBackOffRetry
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.HttpUtils
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeObserver
@@ -167,6 +171,7 @@ class MeetingFragment : Fragment(),
     private lateinit var noVideoOrScreenShareAvailable: TextView
     private lateinit var editTextMessage: EditText
     private lateinit var buttonMute: ImageButton
+    private lateinit var buttonSpeaker: ImageButton
     private lateinit var buttonCamera: ImageButton
     private lateinit var deviceAlertDialogBuilder: AlertDialog.Builder
     private lateinit var additionalOptionsAlertDialogBuilder: AlertDialog.Builder
@@ -191,13 +196,13 @@ class MeetingFragment : Fragment(),
     private lateinit var audioDeviceManager: AudioDeviceManager
 
     companion object {
-        fun newInstance(meetingId: String): MeetingFragment {
+        fun newInstance(meetingId: String, audioVideoConfig: AudioVideoConfiguration): MeetingFragment {
             val fragment = MeetingFragment()
 
-            fragment.arguments =
-                Bundle().apply {
-                    putString(HomeActivity.MEETING_ID_KEY, meetingId)
-                }
+            fragment.arguments = bundleOf(
+                HomeActivity.MEETING_ID_KEY to meetingId,
+                HomeActivity.AUDIO_MODE_KEY to audioVideoConfig.audioMode.value
+            )
             return fragment
         }
     }
@@ -258,7 +263,12 @@ class MeetingFragment : Fragment(),
 
         selectTab(meetingModel.tabIndex)
         setupAudioVideoFacadeObservers()
-        audioVideo.start()
+        val audioMode = arguments?.getInt(HomeActivity.AUDIO_MODE_KEY)?.let { intValue ->
+            AudioMode.from(intValue, defaultAudioMode = AudioMode.Stereo48K)
+        } ?: AudioMode.Stereo48K
+        val audioVideoConfig = AudioVideoConfiguration(audioMode = audioMode)
+        // Start Audio Video
+        audioVideo.start(audioVideoConfig)
         audioVideo.startRemoteVideo()
         return view
     }
@@ -267,6 +277,8 @@ class MeetingFragment : Fragment(),
         buttonMute = view.findViewById(R.id.buttonMute)
         buttonMute.setImageResource(if (meetingModel.isMuted) R.drawable.button_mute_on else R.drawable.button_mute)
         buttonMute.setOnClickListener { toggleMute() }
+
+        buttonSpeaker = view.findViewById(R.id.buttonSpeaker)
 
         buttonCamera = view.findViewById(R.id.buttonCamera)
         buttonCamera.setImageResource(if (meetingModel.isCameraOn) R.drawable.button_camera_on else R.drawable.button_camera)
@@ -599,30 +611,7 @@ class MeetingFragment : Fragment(),
     }
 
     override fun onAttendeesJoined(attendeeInfo: Array<AttendeeInfo>) {
-        uiScope.launch {
-            mutex.withLock {
-                attendeeInfo.forEach { (attendeeId, externalUserId) ->
-                    if (DefaultModality(attendeeId).hasModality(ModalityType.Content) &&
-                        !isSelfAttendee(attendeeId) &&
-                        meetingModel.isSharingContent) {
-                        audioVideo.stopContentShare()
-                        screenShareManager?.stop()
-                        val name = meetingModel.currentRoster[DefaultModality(attendeeId).base()]?.attendeeName ?: ""
-                        notifyHandler("$name took over the screen share")
-                    }
-                    meetingModel.currentRoster.getOrPut(
-                        attendeeId,
-                        {
-                            RosterAttendee(
-                                attendeeId,
-                                getAttendeeName(attendeeId, externalUserId)
-                            )
-                        })
-                }
-
-                rosterAdapter.notifyDataSetChanged()
-            }
-        }
+        onAttendeesJoinedWithStatus(attendeeInfo, AttendeeStatus.Joined)
     }
 
     override fun onAttendeesLeft(attendeeInfo: Array<AttendeeInfo>) {
@@ -716,6 +705,35 @@ class MeetingFragment : Fragment(),
             scoresStr,
             LogLevel.DEBUG
         )
+    }
+
+    private fun onAttendeesJoinedWithStatus(attendeeInfo: Array<AttendeeInfo>, status: AttendeeStatus) {
+        uiScope.launch {
+            mutex.withLock {
+                attendeeInfo.forEach { (attendeeId, externalUserId) ->
+                    if (DefaultModality(attendeeId).hasModality(ModalityType.Content) &&
+                            !isSelfAttendee(attendeeId) &&
+                            meetingModel.isSharingContent) {
+                        audioVideo.stopContentShare()
+                        screenShareManager?.stop()
+                        val name = meetingModel.currentRoster[DefaultModality(attendeeId).base()]?.attendeeName ?: ""
+                        notifyHandler("$name took over the screen share")
+                    }
+                    meetingModel.currentRoster.getOrPut(
+                        attendeeId,
+                        {
+                            RosterAttendee(
+                                attendeeId,
+                                getAttendeeName(attendeeId, externalUserId),
+                                attendeeStatus = status
+                            )
+                        }
+                    )
+                }
+
+                rosterAdapter.notifyDataSetChanged()
+            }
+        }
     }
 
     private fun getAttendeeName(attendeeId: String, externalUserId: String): String {